Moved server-nodejs/lib/* into server-nodejs/db/*
